+++ cnd.makeproject/src/org/netbeans/modules/cnd/makeproject/configurations/AuxConfigurationXMLCodec.java
@@ -86,7 +86,7 @@ class AuxConfigurationXMLCodec extends CommonConfigurationXMLCodec {
     public void start(Attributes atts) throws VersionException {
         String versionString = atts.getValue("version");        // NOI18N
         if (versionString != null) {
-            descriptorVersion = new Integer(versionString).intValue();
+            descriptorVersion = new Integer(versionString);
         }
     }
 
@@ -131,7 +131,7 @@ class AuxConfigurationXMLCodec extends CommonConfigurationXMLCodec {
     @Override
     public void endElement(String element, String currentText) {
         if (element.equals(DEFAULT_CONF_ELEMENT)) {
-            configurationDescriptor.getConfs().setActive(new Integer(currentText).intValue());
+            configurationDescriptor.getConfs().setActive(new Integer(currentText));
         } else if (element.equals(DEVELOPMENT_SERVER_ELEMENT)) {
             if (currentConf instanceof MakeConfiguration) {
                 ExecutionEnvironment env = ExecutionEnvironmentFactory.fromUniqueID(currentText);
@@ -140,7 +140,7 @@ class AuxConfigurationXMLCodec extends CommonConfigurationXMLCodec {
             }
         } else if (element.equals(PLATFORM_ELEMENT)) {
             if (currentConf instanceof MakeConfiguration) {
-                int set = new Integer(currentText).intValue();
+                int set = new Integer(currentText);
                 if (descriptorVersion <= 37 && set == 4) {
                     set = PlatformTypes.PLATFORM_GENERIC;
                 }
